+++ playgrounds/demo/app/pages/components/table.vue
@@ -474,8 +474,8 @@ onMounted(() => {
             onUpdateChecked(checked: boolean) {
               table?.tableApi?.getColumn(column.id)?.toggleVisibility(!!checked)
             },
-            onSelect(e?: Event) {
-              e?.preventDefault()
+            onSelect(e: Event) {
+              e.preventDefault()
             }
           }))"
           :content="{ align: 'end' }"